requirements.txt
fire==0.4.0
importlib-metadata==5.0.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
PyYAML==6.0
six==1.16.0
termcolor==2.1.0
zipp==3.10.0
